requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
Pillow>=9.0.0
Werkzeug>=2.0.0
